requirements.txt
alembic==1.13.1
annotated-types==0.7.0
anyio==4.4.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
async-timeout==4.0.3
asyncpg==0.29.0
boto3==1.34.125
botocore==1.34.125
build==1.2.1
cachetools==5.3.3
certifi==2024.6.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cryptography==42.0.8
dill==0.3.6
dnslib==0.9.24
dnspython==2.6.1
ecdsa==0.19.0
email_validator==2.1.1
fastapi==0.111.0
fastapi-cli==0.0.4
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
idna==3.7
iniconfig==2.0.0
Jinja2==3.1.4
jmespath==1.0.1
localstack==3.5.0
localstack-core==3.5.0
localstack-ext==3.5.0
Mako==1.3.5
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
minio==7.2.7
orjson==3.10.4
packaging==24.1
pluggy==1.5.0
plux==1.10.0
psutil==5.9.8
pyaes==1.6.1
pyasn1==0.6.0
pycparser==2.22
pycryptodome==3.20.0
pydantic==2.7.4
pydantic_core==2.18.4
Pygments==2.18.0
pyproject_hooks==1.1.0
pytest==8.2.2
pytest-asyncio==0.23.7
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.9
PyYAML==6.0.1
requests==2.32.3
rich==13.7.1
rsa==4.9
s3transfer==0.10.1
semver==3.0.2
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.30
starlette==0.37.2
tabulate==0.9.0
tailer==0.4.1
typer==0.12.3
typing_extensions==4.12.2
ujson==5.10.0
urllib3==2.2.1
uvicorn==0.30.1
watchfiles==0.22.0
websockets==12.0
windows-curses==2.3.3
